Actual game graphics
Change placeholders for actual textures, changed train from a cube to a plane since it's just a texture, added rails (I've heard that could be usefull), implemented functions to actually change the next biome (not hard set anymore). I think that's all?...
I'm not putting the textures in the commit. They're quite heavy.